Got sound almost working again

+++ crates/snemcore/src/ssmp/sdsp/voices.rs
@@ -36,7 +36,7 @@ pub struct VoiceRegs {
     pub envelope: i16,
 
     // $X9
-    pub sample_out_high: u8,
+    pub sample_out_high: i16,
 
     // $XA, $XB
     pub ram_a: u8,
@@ -58,7 +58,7 @@ pub struct VoiceRegs {
     pub adsr_stage: ADSRStage,
     pub prev_interpolation_idx: usize,
     pub interpolation_idx: usize,
-    pub brr_sample_buffer: [u16; 12],
+    pub brr_sample_buffer: [i16; 12],
     pub brr_group_addr: u16, // Base address of the BRR sample group (9 bytes)
     pub brr_group_step: usize, // Keeps track of how many sets of 4 BRR samples
     // have been read into the buffer so far from
